Move report structure variant change to export page
